remove old script
